manifest: set memfault to follow master branch
even though we're currently not using it, once we start using it we want
to have the latest version of the sdk. Memfault only pushes release
versions to the master branch, so we should not end up with a
development version of the SDK.

Once used in production, we might want to stick to a specific version of
the SDK, in case breaking changes are introduced.